+++ src/components/level/toolbars/LevelConfig/index.ts
@@ -1,3 +1,4 @@
 export * from "./LevelConfig";
 export * from "./LevelConfigDialog";
+export * from "./LevelDriverConfig";
 export * from "./NoticeSizeLags";